+++ mycodo/utils/github_release_info.py
@@ -1,19 +1,12 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
-try:
-    # For Python 3.0 and later
-    from urllib.request import urlopen
-except ImportError:
-    # Fall back to Python 2's urllib2
-    from urllib2 import urlopen
-
 import argparse
 import json
 import logging
+import sys
+from urllib.request import urlopen
+
 import os
 import re
-import sys
 
 sys.path.append(
     os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
@@ -40,7 +33,7 @@ def json_to_dict(url):
 
 
 def github_releases(major_version):
-    """ Return the tarball URL for the latest version """
+    """ Return the tarball URL for the latest Mycodo release version """
     mycodo_releases = json_to_dict(release_url)
     all_versions = []
     for each_release in mycodo_releases:
@@ -52,8 +45,8 @@ def github_releases(major_version):
 
 def is_latest_installed(major_number):
     """
-    Check if the latest release is installed.
-    Return True if yes, Flase if no.
+    Check if the latest Mycodo release version is installed.
+    Return True if yes, False if no.
     """
     latest_version = return_maj_version_url(True, major_number)
     if latest_version == MYCODO_VERSION:
@@ -63,8 +56,8 @@ def is_latest_installed(major_number):
 
 def sort_reverse_list(versions_unsorted):
     """
-    Sort and reverse a list of strings representing version numbers in
-    the format "x.x.x"
+    Sort and reverse a list of strings representing Mycodo release version
+    numbers in the format "x.x.x"
 
     :return: list of sorted version strings
     :rtype: list
@@ -81,7 +74,7 @@ def sort_reverse_list(versions_unsorted):
 
 
 def return_latest_version_url(version_only):
-    """ Return the tarball URL for the latest version """
+    """ Return the tarball URL for the latest Mycodo release version """
     mycodo_releases = json_to_dict(release_url)
     all_versions = []
     for each_release in mycodo_releases:
@@ -91,10 +84,6 @@ def return_latest_version_url(version_only):
     for each_release in mycodo_releases:
         if (re.match('v.*(\d\.\d\.\d)', each_release['name']) and
                 each_release['name'][1:] == sort_reverse_list(all_versions)[0]):
-            # print("\nLatest Version: {ver}"
-            #       "\nTar URL: {url}".format(ver=each_release['name'],
-            #                                 url=each_release['tarball_url']))
-            # return each_release['name'][1:], each_release['tarball_url']
             if version_only:
                 return each_release['name'][1:]
             else:
@@ -103,7 +92,7 @@ def return_latest_version_url(version_only):
 
 def return_maj_version_url(version_only, major_version):
     """
-    Return the tarball URL for the version with the
+    Return the tarball URL for the Mycodo release version with the
     specified major number
     """
     mycodo_releases = json_to_dict(release_url)
@@ -116,11 +105,6 @@ def return_maj_version_url(version_only, major_version):
     for each_release in mycodo_releases:
         if (re.match('v{maj}.*(\d\.\d)'.format(maj=major_version), each_release['name']) and
                 each_release['name'][1:] == sort_reverse_list(maj_versions)[0]):
-            # print("\nLatest v{maj} Version: {ver}"
-            #       "\nTar URL: {url}".format(maj=major_version,
-            #                                 ver=each_release['name'],
-            #                                 url=each_release['tarball_url']))
-            # return each_release['name'][1:], each_release['tarball_url']
             if version_only:
                 return each_release['name'][1:]
             else:
@@ -129,7 +113,7 @@ def return_maj_version_url(version_only, major_version):
 
 def version_information(version_only, major_version):
     """
-    Print all releases, and specific info about
+    Print all Mycodo releases, and specific info about
     latest and major releases
     """
     mycodo_releases = json_to_dict(release_url)
@@ -142,21 +126,20 @@ def version_information(version_only, major_version):
     print(return_maj_version_url(version_only, major_version))
 
 
-def parseargs(parser):
-    parser.add_argument('-c', '--currentversion', action='store_true',
-                        help='Return the currently-installed version.')
-    parser.add_argument('-i', '--islatest', action='store_true',
-                        help='Return True if the currently-installed '
-                             'version is the latest.')
-    parser.add_argument('-l', '--latest', action='store_true',
-                        help='Return the latest version URL.')
-    parser.add_argument('-m', '--majornumber', type=int,
-                        help='Return the latest version URL with major '
-                             'version'
-                             ' number x in x.y.z.')
-    parser.add_argument('-v', '--version', action='store_true',
-                        help='Return the latest version number.')
-    return parser.parse_args()
+def parseargs(p):
+    p.add_argument('-c', '--currentversion', action='store_true',
+                   help='Return the currently-installed version.')
+    p.add_argument('-i', '--islatest', action='store_true',
+                   help='Return True if the currently-installed '
+                        'version is the latest.')
+    p.add_argument('-l', '--latest', action='store_true',
+                   help='Return the latest version URL.')
+    p.add_argument('-m', '--majornumber', type=int,
+                   help='Return the latest version URL with major '
+                        'version number x in x.y.z.')
+    p.add_argument('-v', '--version', action='store_true',
+                   help='Return the latest version number.')
+    return p.parse_args()
 
 
 if __name__ == "__main__":